+++ CHANGELOG.md
@@ -397,6 +397,9 @@ For instructions on upgrading to newer versions, visit
 
 * \#1135 DateTimes now properly get time zones on derserialization.
 
+* \#685 Attempting to use versioning with embedded documents will now
+  raise a proper error alerting the developer this is not allowed.
+
 ## 2.4.6 (branch: 2.4.0-stable)
 
 ### Resolved Issues